requirements.txt
altgraph==0.17.3
astroid==2.5.6
async-generator==1.10
attrs==23.1.0
autopep8==1.6.0
beautifulsoup4==4.12.0
camelcase==0.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.6
colour==0.1.5
crypto==1.4.1
decouple==0.0.7
dokusan==0.1.0a6
exceptiongroup==1.1.1
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
future==0.18.2
greenlet==2.0.2
h11==0.14.0
idna==3.3
imageio==2.27.0
importlib-metadata==5.2.0
isort==5.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.6.0
lazy_loader==0.2
lxml==4.9.2
MarkupSafe==2.1.1
mccabe==0.6.1
MouseInfo==0.1.3
mysql-connector-python==8.0.30
Naked==0.1.31
networkx==3.1
numpy==1.23.4
opencv-python==4.7.0.72
outcome==1.2.0
packaging==23.0
pefile==2022.5.30
Pillow==9.2.0
protobuf==3.20.1
PyAutoGUI==0.9.53
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.15.0
PyGetWindow==0.0.9
pyinstaller==5.4.1
pyinstaller-hooks-contrib==2022.10
pylint==2.8.2
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-decouple==3.8
python-dotenv==0.21.0
pytweening==1.0.4
PyWavelets==1.4.1
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.28.1
scikit-image==0.20.0
scipy==1.9.1
selenium==4.9.0
shellescape==3.8.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
SQLAlchemy==2.0.7
stripe==5.4.0
tifffile==2023.3.21
tkmacosx==1.0.5
toml==0.10.2
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
urllib3==1.26.11
Werkzeug==2.2.2
wrapt==1.12.1
wsproto==1.2.0
xlwt==1.3.0
zipp==3.11.0
